pyproject.toml
[tool.poetry]
name = "otterdog"
version = "0.3.0.dev"
description = "Tool to manage GitHub organizations and their repositories."
authors = ["Thomas Neidhart <thomas.neidhart@eclipse-foundation.org>"]
readme = "README.md"
packages = [{include = "otterdog"}]

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.38"
jq = "^1.6"
colorama = "^0.4"
mintotp = "^0.3"
requests = "^2.30"
jsonschema = "^4.19"
JSONBender = "^0.9"
importlib_resources = "^5.12"
requests-cache = "^1.1"
gojsonnet = "^0.20"
click = "^8.1"
chevron = "^0.14"
PyNaCl = "^1.5"
jwt = "^1.3"
aiohttp = "^3.8"
aiohttp-client-cache = "^0.9"
aiofiles = "^23.2"
aiosqlite = "^0.19"

[tool.poetry.group.app.dependencies]
flask = "^2.3"
flask-login = "^0.6"
flask-migrate = "^3.1"
flask-dance = "^7.0"
flask-restx = "^1.1"
flask-wtf = "^1.1"
flask-sqlalchemy = "^3.1"
sqlalchemy = {extras = ["mypy"], version = "^2.0"}
WTForms = "^3.0"
email-validator = "^2.0"
python-decouple = "^3.8"
gunicorn = "^21.2"
jinja2 = "^3.1"
Werkzeug = "^2.3"
python-dotenv = "^1.0"
pydantic = "^2.4"
celery = "^5.2"
redis = "^5.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0"
pre-commit = "^3.3"
black = "^23.3"

[tool.poetry.group.test.dependencies]
pytest = "^7.3"

[tool.poetry.group.typing.dependencies]
mypy = "^1.3"
types-colorama = "^0.4"
types-requests = "^2.30"
types-jsonschema = "^4.17"
types-chevron = "^0.14"
types-Flask-Migrate = "^4.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4"
mkdocs-material = "^9.1"
mkdocs-exclude = "^1.0"
jinja2 = "^3.1"

[tool.poetry.scripts]
otterdog = "otterdog.cli:cli"
otterdog-app = "otterdog.app:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.2"
testpaths = ["tests"]
pythonpath = [".", "otterdog"]

[tool.black]
color = true
line-length = 120
target-version = ['py310']
skip-string-normalization = true

[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin"